requirements.txt
weaviate-client==4.20.5
openai==1.59.2
python-dotenv==1.0.1
python-telegram-bot==21.10
pytest==8.3.4
